requirements.txt
GitPython==3.1.41
schedule==1.2.1 